# docs/openapi.yaml
openapi: 3.0.3
info:
  title: 车载视觉感知与人机交互系统 API
  version: 1.0.0
  description: |
    覆盖车牌识别、交警手势识别、车主手势控车、日志监控与告警智能体四个模块。
    Swagger UI：`/api/docs`；原始运行时定义：`/api/openapi.json`。
servers:
  - url: http://{host}:8001
    variables:
      host:
        default: localhost
...
paths:
...
  /api/lpr/recognize:
    post:
      tags: [车牌识别]
      summary: 上传图像识别车牌
      security: [{bearerAuth: []}]
      requestBody: {$ref: '#/components/requestBodies/ImageFile'}
      responses:
        '200': {description: 识别完成, content: {application/json: {schema: {$ref: '#/components/schemas/LPRResponse'}}}}
        '500': {$ref: '#/components/responses/RecognitionFailure'}
  /api/lpr/recognize-video:
    post:
      tags: [车牌识别]
      summary: 上传视频并按间隔抽帧识别
      security: [{bearerAuth: []}]
      parameters:
        - name: interval
          in: query
          description: 采样间隔，单位为帧
          schema: {type: integer, minimum: 1, maximum: 60, default: 15}
      requestBody: {$ref: '#/components/requestBodies/VideoFile'}
      responses:
        '200': {description: 视频处理完成, content: {application/json: {schema: {$ref: '#/components/schemas/LPRVideoResponse'}}}}
        '500': {$ref: '#/components/responses/RecognitionFailure'}
  /api/lpr/history:
    get:
      tags: [车牌识别]
      summary: 查询当前用户及匿名车牌识别历史
      security: [{bearerAuth: []}]
      parameters: [{$ref: '#/components/parameters/Skip'}, {$ref: '#/components/parameters/HistoryLimit'}]
      responses:
        '200': {description: 历史记录, content: {application/json: {schema: {type: array, items: {$ref: '#/components/schemas/LPRHistoryItem'}}}}}
...
components:
  securitySchemes:
    bearerAuth: {type: http, scheme: bearer, bearerFormat: JWT}
  parameters:
...
    Skip: {name: skip, in: query, schema: {type: integer, minimum: 0, default: 0}}
    HistoryLimit: {name: limit, in: query, schema: {type: integer, minimum: 1, default: 20}}
  requestBodies:
    ImageFile:
      required: true
      content: {multipart/form-data: {schema: {type: object, required: [file], properties: {file: {type: string, format: binary, description: 图片或 GIF 文件}}}}}
    VideoFile:
      required: true
      content: {multipart/form-data: {schema: {type: object, required: [file], properties: {file: {type: string, format: binary, description: 视频文件}}}}}
  responses:
...
    RecognitionFailure: {description: 图像/模型处理失败, content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}
...
  schemas:
...
    PlateResult: {type: object, required: [plate_number, plate_color, confidence, bbox], properties: {plate_number: {type: string}, plate_color: {type: string}, confidence: {type: number, format: float}, bbox: {type: array, items: {type: integer}, minItems: 4, maxItems: 4}}}
    LPRResponse: {type: object, required: [plates, plate_count, annotated_image, success], properties: {plates: {type: array, items: {$ref: '#/components/schemas/PlateResult'}}, plate_count: {type: integer}, annotated_image: {type: string, description: JPEG 的 Base64 编码}, success: {type: boolean}, record_id: {type: integer, nullable: true}}}
    LPRVideoResponse: {type: object, properties: {frame_count: {type: integer}, results: {type: array, items: {type: object, additionalProperties: true}}}}
    LPRHistoryItem: {type: object, properties: {id: {type: integer}, source_type: {type: string}, plate_count: {type: integer}, annotated_image: {type: string, nullable: true}, created_at: {type: string, format: date-time}}}
...
    Error: {type: object, required: [detail], properties: {detail: {type: string, description: FastAPI 当前实现的错误消息}}}
...